lab6/keycloakuniversityportal/requirements.txt
cryptography==46.0.3
dotenv==0.9.9
Flask==3.0.3
PyJWT==2.9.0
